% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt.R
\name{check_cwt_recoded}
\alias{check_cwt_recoded}
\title{Check CWT recoded}
\usage{
check_cwt_recoded(conn, proj)
}
\arguments{
\item{conn}{A database connection object.}

\item{proj}{A character vector of project codes.}
}
\value{
Invisibly returns a list with `aligned_missing_rows` and `ann_missing_rows`.
}
\description{
Checks if all variables from the aligned CWT and annotation
  sheet are present in the final CWT to recode.
}
\keyword{internal}
