% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{clean_varnames}
\alias{clean_varnames}
\title{Clean Variable Names in a Dataset}
\usage{
clean_varnames(data)
}
\arguments{
\item{data}{A data frame whose variable names should be cleaned}
}
\value{
A data frame with cleaned variable names
}
\description{
Converts all variable names in a dataset to a clean format by:
- Converting to lowercase
- Replacing spaces and dots with underscores
- Removing special characters (keeping only letters, numbers, and underscores)
}
\examples{
\dontrun{
df <- data.frame("First Name" = 1:3, "Last.Name" = letters[1:3],
                 "Age (years)" = c(25, 30, 35))
clean_df <- clean_varnames(df)
names(clean_df)  # "first_name" "last_name" "age_years"
}

}
