% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.R
\name{load_cwt}
\alias{load_cwt}
\title{Load crosswalk tables from disk}
\usage{
load_cwt(proj, status = "appended")
}
\arguments{
\item{proj}{Survey Project identifier.}

\item{status}{CWT version to load. Defaults to `"appended"`.}
}
\value{
For `"aligned"` and `"original"`, returns a single CWT.

For `"appended"` and `"recoding"`, loads all files and assigns them to the
global environment.
}
\description{
Load crosswalk tables (CWTs) for a project from disk.
}
\details{
Reads files from project-specific directories under
`Sys.getenv("ROOT_DIR")`.
}
\examples{
\dontrun{
cwt <- load_cwt("ess", "aligned")
load_cwt("ess", "appended")
}
}
\seealso{
[bind_cwts()], [write_cwt()]

Other cwt-helpers: 
\code{\link{load_annotations}()},
\code{\link{load_survey_list}()},
\code{\link{load_var_validations}()}
}
\concept{cwt-helpers}
\keyword{internal}
