% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt.R
\name{bind_cwts}
\alias{bind_cwts}
\title{Combine crosswalk tables stored as separate files}
\usage{
bind_cwts(proj, status, indicator, timestamp = timestamp())
}
\arguments{
\item{proj}{A single project code, such as \code{"ess"} or \code{"eb"}.}

\item{status}{Which CWT version to load. Must be one of
\code{"appended"} or \code{"recoded"}.}

\item{indicator}{A character vector of indicator groups to load, such as
\code{c("swd", "controls")}.}

\item{timestamp}{Uses the current `sysdate()` timestamp as default. Otherwise supply timestamp for date of CWT creation.}
}
\value{
A data frame formed by row-binding the requested CWT files which - Columns are not modified.
}
\description{
Read and row-bind crosswalk table (CWT) files for one or more indicators
from a project's \code{appended} or \code{recoded} directory.
}
\details{
For each indicator, the function looks for a file named
\code{<indicator>.<ext>} and reads the first match in the order
\code{.xlsx}, \code{.odt}, \code{.csv}.

The \code{ROOT_DIR} environment variable must point to the project root.
}
\examples{
\dontshow{if (Sys.getenv("ROOT_DIR") != "") withAutoprint(\{ # examplesIf}
bind_cwts(
  proj = "ess",
  status = "appended",
  indicator = c("swd", "controls")
)
\dontshow{\}) # examplesIf}
}
\references{
Kołczyńska, M. (2022). Combining multiple survey sources: A reproducible
workflow and toolbox for survey data harmonization. \emph{Methodological
Innovations}, 15(1), 62--72. \doi{10.1177/20597991221077923}
}
\seealso{
* [write_cwt()] saves a CWT to the directory.
 * [build_cwt()] creates a CWT from scratch.
 * [append_cwt()] binds missing variables to an existing CWT.

Other cwt: 
\code{\link{append_cwt}()},
\code{\link{build_cwt}()},
\code{\link{check_cwt_recoded}()},
\code{\link{write_cwt}()}
}
\concept{cwt}
