% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt_create.R
\name{create_cwt}
\alias{create_cwt}
\title{Create a Codebook with Value Labels (CWT)}
\usage{
create_cwt(survey_data, variable_map, verbose = TRUE)
}
\arguments{
\item{survey_list}{A named list of survey data frames, keyed by study_wave}

\item{vars_df}{A data frame with columns \code{src_var}, \code{target_var},
and \code{study_wave}}
}
\value{
A CWT data frame with one row per value per variable per wave
}
\description{
Create a Codebook with Value Labels (CWT)
}
